AmberFlux/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google-genai